1.16.5-0.0.18: Don't trust Double.MIN_VALUE (temperature fix)

+++ common/common-config/src/main/java/com/pg85/otg/config/standard/BiomeStandardValues.java
@@ -109,8 +109,8 @@ public class BiomeStandardValues extends Settings
 		MAX_AVERAGE_DEPTH = doubleSetting("MaxAverageDepth", 0, -1000, 1000),
 		MOOD_OFFSET = doubleSetting("MoodOffset", 2.0, 0, Double.MAX_VALUE),
 		ADDITIONS_TICK_CHANCE = doubleSetting("AdditionsTickChance", 0, 0, Double.MAX_VALUE),
-		TEMPLATE_FOR_BIOME_MIN_TEMP = doubleSetting("TemplateForBiomeMinTemp", 0, Double.MIN_VALUE, Double.MAX_VALUE),
-		TEMPLATE_FOR_BIOME_MAX_TEMP = doubleSetting("TemplateForBiomeMaxTemp", 0, Double.MIN_VALUE, Double.MAX_VALUE)
+		TEMPLATE_FOR_BIOME_MIN_TEMP = doubleSetting("TemplateForBiomeMinTemp", 0, Integer.MIN_VALUE, Double.MAX_VALUE),
+		TEMPLATE_FOR_BIOME_MAX_TEMP = doubleSetting("TemplateForBiomeMaxTemp", 0, Integer.MIN_VALUE, Double.MAX_VALUE)
 	;
 
 	public static final Setting<LocalMaterialData>
